// src/enums/ws.enum.ts
export enum WsMessageName {
  Allocate = "ALLOCATE",
  CancelEngagement = "CANCEL_ENGAGEMENT",
  ConfirmPosition = "CONFIRM_POSITION",
  GetMissionsList = "GET_MISSIONS_LIST",
  GetDb = "GET_DB",
  LoadMission = "LOAD_MISSION",
  LosRequest = "LOS_REQUEST",
  SaveEntity = "SAVE_ENTITY",
  UpdateEntity = "UPDATE_ENTITY",
  EntityCreated = "ENTITY_CREATED",
  Ping = "PING",
  SaveMission = "SAVE_MISSION",
  SetPosition = "SET_POSITION",
  SetRadarParams = "SET_RADAR_PARAMS",
  SetTabooZone = "SET_TABOO_ZONE",
  SetTargetInfo = "SET_TARGET_INFO",
  SystemMode = "SYSTEM_MODE",
  EntityAdded = "ENTITY_ADDED",
  EntityDeleted = "DELETE_ENTITY",
  DeleteMission = "DELETE_MISSION",
  EntityPosition = "ENTITY_POSITION",
  EntityUpdated = "ENTITY_UPDATED",
  LosSector = "LOS_SECTOR",
  LoadResult = "LOAD_RESULT",
  Error = "ERROR",
  MissionsList = "MISSIONS_LIST",
  MissionsListUpdate = "MISSIONS_LIST_UPDATE",
  MissionData = "MISSION_DATA",
  MissionDataUpdate = "MISSION_DATA_UPDATE",
  Position = "TMAPS_PARAMS",
  RadarUpdate = "RADAR_UPDATE",
  SaveResult = "SAVE_RESULT",
  RecommendAssignment = "RECOMMEND_ASSIGNMENT",
  TargetUpdate = "TARGET_UPDATE",
  LosResult = "LOS_RESULT",
  LosUpdate = "LOS_UPDATE",
  TargetsData = "TARGETS_DATA",
  RadarStatus = "RADAR_STATUS",
  RadarParamsUpdate = "RADAR_PARAMS_UPDATE",
  RadarBitStatusUpdate = "RADAR_BIT_STATUS_UPDATE",
  ConfirmUpdate = "CONFIRM_UPDATE",
  SystemStatusM = "SYSTEM_STATUS_M",
  RadarParams = "RADAR_PARAMS",
  RadarBitStatus = "RADAR_BIT_STATUS",
  TargetAssigned = "TARGET_Assigned",
  TargetLock = "TARGET_LOCK",
  SystemStatus = "SYSTEM_STATUS",
  TargetDestroyed = "TARGET_DESTROYED",
  GunBitStatus = "GUN_BIT_STATUS",
  OdoCaliFinished = "ODO_CALI_FINISHED",
  TmapsBitStatus = "TMAPS_BIT_STATUS",
  StartRealing = "START_REALIGN",
  StartOdoCali = "START_ODO_CALI",
  SetPosType = "SET_POS_TYPE",
  GpsIntegration = "GPS_INTEGRATION",
  Gun_Params = "GUN_PARAMS",
}